backend/requirements.txt
# Flask Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.23

# Validation & Serialization
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Password Hashing
bcrypt==4.1.2

# Environment Variables
python-dotenv==1.0.0

# IP Address Validation
ipaddress==1.0.23

# Date/Time Utilities
python-dateutil==2.8.2

# Testing (for later phases)
pytest==7.4.3
pytest-flask==1.3.0

# Production Server
gunicorn==21.2.0
